refactor: move ValueObjects de 'File' para o pacote "org.osnormais.storage.api.domain.file.valueobject"
