+++ domain/src/main/java/org/osnormais/storage/api/domain/file/valueobject/ParallelChunkLimit.java
@@ -1,16 +1,16 @@
-package org.osnormais.storage.api.domain.file;
+package org.osnormais.storage.api.domain.file.valueobject;
 
 import org.osnormais.storage.api.domain.ValueObject;
 import org.osnormais.storage.api.domain.validation.ValidationError;
 import org.osnormais.storage.api.domain.validation.handler.ValidationHandler;
 
-public record Size(long bytes) implements ValueObject {
+public record ParallelChunkLimit(int value) implements ValueObject {
 
     @Override
     public void validate(final ValidationHandler handler) {
 
-        if (bytes <= 0)
-            handler.append(ValidationError.with("bytes must be greater than 0"));
+        if (value < 1)
+            handler.append(ValidationError.with("'value' must be greater than or equal to 1"));
 
     }
