+++ domain/src/main/java/org/osnormais/storage/api/domain/file/valueobject/Size.java
@@ -1,16 +1,16 @@
-package org.osnormais.storage.api.domain.file;
+package org.osnormais.storage.api.domain.file.valueobject;
 
 import org.osnormais.storage.api.domain.ValueObject;
 import org.osnormais.storage.api.domain.validation.ValidationError;
 import org.osnormais.storage.api.domain.validation.handler.ValidationHandler;
 
-public record ThroughputLimit(long bytesPerSecond) implements ValueObject {
+public record Size(long bytes) implements ValueObject {
 
     @Override
     public void validate(final ValidationHandler handler) {
 
-        if (bytesPerSecond < 0)
-            handler.append(ValidationError.with("bytesPerSecond must be greater than or equal to 0"));
+        if (bytes <= 0)
+            handler.append(ValidationError.with("bytes must be greater than 0"));
 
     }
