+++ domain/src/main/java/org/osnormais/storage/api/domain/file/valueobject/ThroughputLimit.java
@@ -1,16 +1,16 @@
-package org.osnormais.storage.api.domain.file;
+package org.osnormais.storage.api.domain.file.valueobject;
 
 import org.osnormais.storage.api.domain.ValueObject;
 import org.osnormais.storage.api.domain.validation.ValidationError;
 import org.osnormais.storage.api.domain.validation.handler.ValidationHandler;
 
-public record ParallelChunkLimit(int value) implements ValueObject {
+public record ThroughputLimit(long bytesPerSecond) implements ValueObject {
 
     @Override
     public void validate(final ValidationHandler handler) {
 
-        if (value < 1)
-            handler.append(ValidationError.with("'value' must be greater than or equal to 1"));
+        if (bytesPerSecond < 0)
+            handler.append(ValidationError.with("bytesPerSecond must be greater than or equal to 0"));
 
     }
